requirements.txt
pyautogui==0.9.52
Pillow==8.3.2
opencv-python==4.5.3.56
pytesseract==0.3.8
keyboard==0.13.5
tkinter==8.6  
